Add bulk quotation creation and refactor DTOs
Introduces an endpoint to create multiple quotations for an order, updates QuotationRequestDTO and related model fields to use Integer for supplierId, removes orderId from ItemResponseDTO and QuotationResponseDTO, and adds existence checks for quotation index. Also updates API documentation and repository methods to support new functionality.

+++ src/main/java/de/hs_esslingen/besy/dtos/request/QuotationRequestDTO.java
@@ -14,5 +14,5 @@ public class QuotationRequestDTO implements Serializable {
     Short index;
     LocalDate quoteDate;
     BigDecimal price;
-    Long supplierId;
+    Integer supplierId;
 }
